+++ betfairlightweight/baseclient.py
@@ -1,9 +1,16 @@
+import os
+import time
 import requests
 import collections
-import datetime
-import os
 
 from .exceptions import PasswordError, AppKeyError, CertsError
+
+IDENTITY = "https://identitysso.betfair{tld}/api/"
+IDENTITY_CERT = "https://identitysso-cert.betfair{tld}/api/"
+API = "https://api.betfair.com/exchange/"
+NAVIGATION = (
+    "https://api.betfair{tld}/exchange/betting/rest/v1/{locale}/navigation/menu.json"
+)
 
 
 class BaseClient:
@@ -12,29 +19,31 @@ class BaseClient:
     """
 
     IDENTITY_URLS = collections.defaultdict(
-        lambda: "https://identitysso.betfair.com/api/",
-        spain="https://identitysso.betfair.es/api/",
-        italy="https://identitysso.betfair.it/api/",
-        romania="https://identitysso.betfair.ro/api/",
-        sweden="https://identitysso.betfair.se/api/",
-        australia="https://identitysso.betfair.au/api/",
+        lambda: IDENTITY.format(tld=".com"),
+        spain=IDENTITY.format(tld=".es"),
+        italy=IDENTITY.format(tld=".it"),
+        romania=IDENTITY.format(tld=".ro"),
+        sweden=IDENTITY.format(tld=".se"),
+        australia=IDENTITY.format(tld=".com.au"),
     )
 
     IDENTITY_CERT_URLS = collections.defaultdict(
-        lambda: "https://identitysso-cert.betfair.com/api/",
-        spain="https://identitysso-cert.betfair.es/api/",
-        italy="https://identitysso-cert.betfair.it/api/",
-        romania="https://identitysso-cert.betfair.ro/api/",
-        sweden="https://identitysso-cert.betfair.se/api/",
+        lambda: IDENTITY_CERT.format(tld=".com"),
+        spain=IDENTITY_CERT.format(tld=".es"),
+        italy=IDENTITY_CERT.format(tld=".it"),
+        romania=IDENTITY_CERT.format(tld=".ro"),
+        sweden=IDENTITY_CERT.format(tld=".se"),
     )
 
-    API_URLS = collections.defaultdict(lambda: "https://api.betfair.com/exchange/")
+    API_URLS = collections.defaultdict(lambda: API)
 
     NAVIGATION_URLS = collections.defaultdict(
-        lambda: "https://api.betfair.com/exchange/betting/rest/v1/en/navigation/menu.json",
-        spain="https://api.betfair.es/exchange/betting/rest/v1/en/navigation/menu.json",
-        italy="https://api.betfair.it/exchange/betting/rest/v1/en/navigation/menu.json",
+        lambda: NAVIGATION.format(tld=".com", locale="en"),
+        spain=NAVIGATION.format(tld=".es", locale="es"),
+        italy=NAVIGATION.format(tld=".it", locale="it"),
     )
+
+    SESSION_TIMEOUT = collections.defaultdict(lambda: 8 * 60 * 60, italy=20 * 60)
 
     def __init__(
         self,
@@ -50,11 +59,11 @@ class BaseClient:
         Creates base client for API operations.
 
         :param str username: Betfair username
-        :param str password: Password for supplied username, if None will look in .bashprofile
+        :param str password: Betfair password for supplied username, if None will look in .bashprofile
         :param str app_key: App Key for account, if None will look in .bashprofile
-        :param str certs: Directory for certificates, if None will look in /certs/
-        :param str locale: Exchange to be used, defaults to UK for login and global for api
-        :param list cert_files: Certificate and key files. If None will look in `certs`
+        :param str certs: Directory for certificates, if None will look in /certs
+        :param str locale: Exchange to be used, defaults to international (.com) exchange
+        :param list cert_files: Certificate and key files. If None will use `self.certs`
         :param bool lightweight: If True endpoints will return dict not a resource (22x faster)
         """
         self.username = username
@@ -72,6 +81,7 @@ class BaseClient:
         self.identity_cert_uri = self.IDENTITY_CERT_URLS[locale]
         self.api_uri = self.API_URLS[locale]
         self.navigation_uri = self.NAVIGATION_URLS[locale]
+        self.session_timeout = self.SESSION_TIMEOUT[locale]
 
         self.get_password()
         self.get_app_key()
@@ -83,12 +93,12 @@ class BaseClient:
         :param str session_token: Session token from request.
         """
         self.session_token = session_token
-        self._login_time = datetime.datetime.now()
+        self._login_time = time.time()
 
     def get_password(self) -> str:
         """
         If password is not provided will look in environment variables
-        for username+'password'.
+        for self.username+'password'.
         """
         if self.password is None:
             if os.environ.get(self.username + "password"):
@@ -120,11 +130,10 @@ class BaseClient:
     def session_expired(self) -> bool:
         """
         Returns True if login_time not set or seconds since
-        login time is greater than 200 mins.
+        login time is greater half session timeout.
         """
-        if (
-            not self._login_time
-            or (datetime.datetime.now() - self._login_time).total_seconds() > 12000
+        if not self._login_time or time.time() - self._login_time > (
+            self.session_timeout / 2
         ):
             return True
         else:
@@ -146,8 +155,8 @@ class BaseClient:
         ssl_path = os.path.join(os.pardir, certs)
         try:
             cert_path = os.listdir(ssl_path)
-        except FileNotFoundError:
-            raise CertsError(certs)
+        except FileNotFoundError as e:
+            raise CertsError(str(e))
 
         cert = None
         key = None
@@ -158,7 +167,7 @@ class BaseClient:
             elif ext == ".key":
                 key = os.path.join(ssl_path, file)
         if cert is None or key is None:
-            raise CertsError(certs)
+            raise CertsError("Certificates not found in directory: '%s'" % ssl_path)
         return [cert, key]
 
     @property
@@ -167,6 +176,7 @@ class BaseClient:
             "Accept": "application/json",
             "X-Application": self.app_key,
             "content-type": "application/x-www-form-urlencoded",
+            "User-Agent": "betfairlightweight",
         }
 
     @property
@@ -176,6 +186,7 @@ class BaseClient:
             "X-Application": self.app_key,
             "X-Authentication": self.session_token,
             "content-type": "application/x-www-form-urlencoded",
+            "User-Agent": "betfairlightweight",
         }
 
     @property