+++ betfairlightweight/metadata.py
@@ -61,7 +61,7 @@ Currency Parameters
 currency_parameters = {
     "GBP": {"min_bet_size": 2, "min_bet_payout": 10, "min_bsp_liability": 10},
     "EUR": {"min_bet_size": 2, "min_bet_payout": 20, "min_bsp_liability": 20},
-    "USD": {"min_bet_size": 4, "min_bet_payout": 20, "min_bsp_liability": 20},
+    "USD": {"min_bet_size": 3, "min_bet_payout": 20, "min_bsp_liability": 20},
     "HKD": {"min_bet_size": 25, "min_bet_payout": 125, "min_bsp_liability": 125},
     "AUD": {"min_bet_size": 5, "min_bet_payout": 30, "min_bsp_liability": 30},
     "CAD": {"min_bet_size": 6, "min_bet_payout": 30, "min_bsp_liability": 30},
@@ -69,9 +69,6 @@ currency_parameters = {
     "NOK": {"min_bet_size": 30, "min_bet_payout": 150, "min_bsp_liability": 150},
     "SEK": {"min_bet_size": 30, "min_bet_payout": 150, "min_bsp_liability": 150},
     "SGD": {"min_bet_size": 6, "min_bet_payout": 30, "min_bsp_liability": 30},
-    "RON": {
-        "min_bet_size": 10,
-        "min_bet_payout": 50,
-        "min_bsp_liability": (10, 50),  # (back, lay)
-    },
+    "RON": {"min_bet_size": 10, "min_bet_payout": 50, "min_bsp_liability": 50},
+    "BRL": {"min_bet_size": 10, "min_bet_payout": 50, "min_bsp_liability": 50},
 }