+++ betfairlightweight/streaming/betfairstream.py
@@ -1,5 +1,6 @@
 import socket
 import ssl
+import logging
 import datetime
 import collections
 from typing import Optional
@@ -7,6 +8,8 @@ from typing import Optional
 from ..exceptions import SocketError, ListenerError
 from ..compat import json
 from .listener import BaseListener
+
+logger = logging.getLogger(__name__)
 
 
 class BetfairStream:
@@ -31,8 +34,7 @@ class BetfairStream:
         session_token: str,
         timeout: float,
         buffer_size: int,
-        description: str,
-        host: str,
+        host: Optional[str],
     ):
         self._unique_id = unique_id
         self.listener = listener
@@ -40,7 +42,6 @@ class BetfairStream:
         self.session_token = session_token
         self.timeout = timeout
         self.buffer_size = buffer_size
-        self.description = description
         self.host = self.HOSTS[host]
         self.receive_count = 0
         self.datetime_last_received = None
@@ -259,6 +260,9 @@ class BetfairStream:
             self._connect()
             self.authenticate()
         message_dumped = json.dumps(message) + self.__CRLF
+        logger.debug(
+            "[Subscription: %s] Sending: %s" % (self._unique_id, repr(message_dumped))
+        )
         try:
             self._socket.send(message_dumped.encode())
         except (socket.timeout, socket.error) as e: